# front/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "E-commerce Platform API",
    "version": "1.0.0",
    "description": "Complete e-commerce platform with comprehensive product management, order processing, payment integration, and user management",
    "contact": {
      "name": "API Support",
      "email": "support@ecommerce-platform.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://api.ecommerce-platform.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token for authentication. Get token from login endpoint and use format: Bearer <token>",
        "example": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for service-to-service authentication"
      }
    },
...
  },
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/admin/payments": {
      "get": {
        "summary": "Get all payments with filtering",
        "tags": [
          "Admin - Payments"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "description": "Page number"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            },
            "description": "Items per page"
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "SUCCESS",
                "FAILED"
              ]
            },
            "description": "Payment status filter"
          },
          {
            "in": "query",
            "name": "gateway",
            "schema": {
              "type": "string",
              "enum": [
                "ZARINPAL",
                "STRIPE",
                "PAYIR"
              ]
            },
            "description": "Payment gateway filter"
          },
          {
            "in": "query",
            "name": "orderId",
            "schema": {
              "type": "integer"
            },
            "description": "Filter by order ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Payments retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Admin access required"
          }
        }
      }
    },
...
  }
}